+++ src/Mod/BIM/TestArch.py
@@ -28,7 +28,7 @@ from bimtests.TestArchSpace import TestArchSpace
 from bimtests.TestArchWall import TestArchWall
 from bimtests.TestArchBuildingPart import TestArchBuildingPart
 from bimtests.TestArchAxis import TestArchAxis
-from bimtests.TestArch import TestArch
+from bimtests.TestArchStructure import TestArchStructure
 from bimtests.TestArchMaterial import TestArchMaterial
 from bimtests.TestArchPanel import TestArchPanel
 from bimtests.TestArchWindow import TestArchWindow
