+++ src/Mod/BIM/bimtests/TestArchFrame.py
@@ -23,15 +23,21 @@
 # ***************************************************************************
 
 import Arch
+import Draft
+import FreeCAD as App
 from bimtests import TestArchBase
 
 class TestArchFrame(TestArchBase.TestArchBase):
 
     def test_makeFrame(self):
         """Test the makeFrame function."""
-        operation = "Testing makeFrame..."
-        self.printTestMessage(operation)
 
         obj = Arch.makeFrame(None, None)
         self.assertIsNotNone(obj, "makeFrame failed to create an object")
         self.assertEqual(obj.Label, "Frame", "Incorrect default label for Frame")
+
+    def testFrame(self):
+        l=Draft.makeLine(App.Vector(0,0,0),App.Vector(-2,0,0))
+        p = Draft.makeRectangle(length=.5,height=.5)
+        f = Arch.makeFrame(l,p)
+        self.assertTrue(f,"Arch Frame failed")
